requirements.txt
aioredis==1.3.1
amqp==5.0.6
apns2==0.7.2
arabic-reshaper==2.1.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
Django==3.2.3
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-cron==0.5.1
django-crontab==0.7.1
django-easy-pdf==0.1.1
django-push-notifications==2.0.0
django-rest-swagger==2.2.0
django-simple-history==3.0.0
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fcm-django==0.3.10
flower==0.9.7
future==0.18.2
h2==4.0.0
hiredis==2.0.0
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
http-ece==1.1.0
humanize==3.6.0
hyper==0.7.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.3.1
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
msgpack==1.0.2
openapi-codec==1.3.2
pdfkit==0.6.1
Pillow==8.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.5.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywebpush==1.13.0
redis==3.5.3
reportlab==3.5.67
requests==2.25.1
service-identity==21.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
stripe==2.57.0
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
xhtml2pdf==0.2.5
zipp==3.4.1
zope.interface==5.4.0
